feat: final implementation adding the verification to the CollaboratorRole and creating a new ResponseDTO

+++ domain/src/main/java/com/salespilot/api/domain/entity/Collaborator.java
@@ -27,7 +27,7 @@ public class Collaborator {
 
     private Integer averageFeeling;
 
-    private Integer totalMeetings;
+    private Long totalMeetings;
 
     private List<Meeting> meetings;
